+++ digiplan/static/js/sliders.js
@@ -119,6 +119,7 @@ PubSub.subscribe(
 PubSub.subscribe(eventTopics.PV_CONTROL_ACTIVATED, showPVLayers);
 PubSub.subscribe(eventTopics.PV_CONTROL_ACTIVATED, highlightPVMapControls);
 PubSub.subscribe(eventTopics.PV_ROOF_CONTROL_ACTIVATED, showPVRoofLayers);
+PubSub.subscribe(eventTopics.WIND_CONTROL_ACTIVATED, updateWindSelection);
 PubSub.subscribe(eventTopics.WIND_CONTROL_ACTIVATED, showWindLayers);
 
 // Subscriber Functions
@@ -306,7 +307,7 @@ function createPercentagesOfPowerSources(msg) {
 }
 
 /* when the other forms get Status Quo marks, there needs to be an iteration over the forms! (line 117)*/
-function updateSliderMarks(msg) {
+export function updateSliderMarks(msg) {
   for (let [slider_name, slider_marks] of Object.entries(
     store.cold.slider_marks,
   )) {
@@ -404,6 +405,23 @@ function showWindLayers(msg) {
   return logMessage(msg);
 }
 
+function updateWindSelection(msg) {
+  const windInput = document.getElementById("id_wind_year");
+  const currentWindTab = document
+    .getElementById("windTab")
+    .getElementsByClassName("active")[0].id;
+  if (currentWindTab === "windPastTab") {
+    windInput.value = "wind_2018";
+  } else if (currentWindTab === "windPresentTab") {
+    windInput.value = "wind_2024";
+  } else if (currentWindTab === "windFutureTab") {
+    windInput.value = "wind_2027";
+  } else {
+    throw Error(`Unknown wind tab '${currentWindTab}' found.`);
+  }
+  return logMessage(msg);
+}
+
 export function hidePotentialLayers(msg) {
   for (const layer of potentialPVLayers
     .concat(potentialPVRoofLayers)
@@ -424,22 +442,21 @@ function highlightPVMapControls(msg) {
 
 function adaptDetailKeyResults(msg, data) {
   const slider_id = data.input[0].id;
-  let technology;
+  let wind_year;
   let target;
   let url_data = {};
 
   if (slider_id === "id_s_w_6") {
-    technology = "wind_2024";
+    wind_year = "wind_2024";
     url_data.id_s_w_6 = data.from;
     target = "wind_key_results_2024";
   } else if (slider_id === "id_s_w_7") {
-    technology = "wind_2027";
+    wind_year = "wind_2027";
     url_data.id_s_w_7 = data.from;
     target = "wind_key_results_2027";
   } else if (
     ["id_s_pv_ff_3", "id_s_pv_ff_4", "id_s_pv_ff_5"].includes(slider_id)
   ) {
-    technology = "pv_ground";
     url_data.id_s_pv_ff_3 =
       $("#id_s_pv_ff_3").data("ionRangeSlider").result.from;
     url_data.id_s_pv_ff_4 =
@@ -448,7 +465,6 @@ function adaptDetailKeyResults(msg, data) {
       $("#id_s_pv_ff_5").data("ionRangeSlider").result.from;
     target = "pv_ground_key_results";
   } else if (slider_id === "id_s_pv_d_3") {
-    technology = "pv_roof";
     url_data.id_s_pv_d_3 = data.from;
     target = "pv_roof_key_results";
   } else {
@@ -456,7 +472,10 @@ function adaptDetailKeyResults(msg, data) {
   }
 
   const query = new URLSearchParams(url_data).toString();
-  let url = `/detail_key_results?technology=${technology}&${query}`;
+  let url = `/detail_key_results?${query}`;
+  if (wind_year !== undefined) {
+    url += "&wind_year=" + wind_year;
+  }
   fetch(url)
     .then((response) => {
       // Check if the response is successful